Initial quest mechanic
Game manager script is added to control and manage the quests.

Collectable objects are added and their order in layer is set to 1, same as the player object.

Added text object to tell the collect what to collect.

Added a sound to play when player touches the correct object.

2D trigger doesn't work so far.

Still dialouges related to quests to be added.